mount_repo_cli()
So this is part of the munkiimport code. I think the return without a statement in line 5 is where my problem comes from. A blank return just invokes None from what I've found, which effectively means nothing, and I think means no matter what the status of the local repo is, that function continues moving and moves to mounting the repo. A kludge fix would be to have line five set the WE_MOUNTED_THE_REPO variable to true, but since they made that a global variable, that may break something else. Does it need to be return True? Would that stop the function from finishing? So I've figured out from setting up a couple of different setups on my various macs that the issue only shows itself if you are attempting to store the repo somewhere other than the /Users/Shared folder if using just the regular apache web server, or directly inside the default web data folder. If using server, I can only get it to work without showing the issue is I put the repo literally inside the default website folder, not even a symbolic link to the /Users/Shared will work in this scenario. And of course, if you're using a share for the repo, the issue would never occur. Thoughts?